Add first test to parse Statistics interface #28
