--- src/app/services/statistics/process-statistics.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class ProcessStatisticsService {
-
-  constructor() { }
-
-}
